+++ packages/expo-blur/android/src/main/java/expo/modules/blur/ExpoBlurView.kt
@@ -1,5 +1,6 @@
 package expo.modules.blur
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Build
 import android.view.ViewGroup
@@ -10,10 +11,11 @@ import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.exception.Exceptions
 import expo.modules.kotlin.views.ExpoView
 
+@SuppressLint("ViewConstructor")
 class ExpoBlurView(context: Context, appContext: AppContext) : ExpoView(context, appContext) {
   private var blurReduction = 4f
   private var blurRadius = 50f
-  internal var tint = "default"
+  internal var tint: TintStyle = TintStyle.DEFAULT
 
   private val blurView = BlurView(context).also {
     it.layoutParams = LayoutParams(LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT)
@@ -47,7 +49,7 @@ class ExpoBlurView(context: Context, appContext: AppContext) : ExpoView(context,
   }
 
   fun applyTint() {
-    blurView.setOverlayColor(tint.toColorInt(blurRadius))
+    blurView.setOverlayColor(tint.toBlurEffect(blurRadius))
     blurView.invalidate()
   }
 }